Move methods to core_actions class and add actions in transactions controller
