+++ controller/ipn_listener.php
@@ -16,7 +16,6 @@ namespace skouat\ppde\controller;
 use phpbb\config\config;
 use phpbb\event\dispatcher_interface;
 use phpbb\language\language;
-use phpbb\path_helper;
 use phpbb\request\request;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -154,8 +153,6 @@ class ipn_listener
 	protected $container;
 	protected $dispatcher;
 	protected $language;
-	protected $path_helper;
-	protected $php_ext;
 	protected $ppde_actions;
 	protected $ppde_controller_main;
 	protected $ppde_controller_transactions_admin;
@@ -175,15 +172,9 @@ class ipn_listener
 	 */
 	private $donor_is_member = false;
 	/**
-	 * @var array|boolean
+	 * @var array
 	 */
 	private $payer_data;
-	/**
-	 * phpBB root path
-	 *
-	 * @var string
-	 */
-	private $root_path;
 	/**
 	 * Data from PayPal transaction
 	 *
@@ -216,7 +207,6 @@ class ipn_listener
 	 * @param config                        $config                             Config object
 	 * @param ContainerInterface            $container                          Service container interface
 	 * @param language                      $language                           Language user object
-	 * @param path_helper                   $path_helper                        Path helper object
 	 * @param core_actions                  $ppde_actions                       PPDE actions object
 	 * @param main_controller               $ppde_controller_main               Main controller object
 	 * @param admin_transactions_controller $ppde_controller_transactions_admin Admin transactions controller object
@@ -224,26 +214,21 @@ class ipn_listener
 	 * @param ipn_paypal                    $ppde_ipn_paypal                    IPN PayPal
 	 * @param request                       $request                            Request object
 	 * @param dispatcher_interface          $dispatcher                         Dispatcher object
-	 * @param string                        $php_ext                            phpEx
 	 *
 	 * @access public
 	 */
-	public function __construct(config $config, ContainerInterface $container, language $language, path_helper $path_helper, core_actions $ppde_actions, main_controller $ppde_controller_main, admin_transactions_controller $ppde_controller_transactions_admin, ipn_log $ppde_ipn_log, ipn_paypal $ppde_ipn_paypal, request $request, dispatcher_interface $dispatcher, $php_ext)
+	public function __construct(config $config, ContainerInterface $container, language $language, core_actions $ppde_actions, main_controller $ppde_controller_main, admin_transactions_controller $ppde_controller_transactions_admin, ipn_log $ppde_ipn_log, ipn_paypal $ppde_ipn_paypal, request $request, dispatcher_interface $dispatcher)
 	{
 		$this->config = $config;
 		$this->container = $container;
 		$this->dispatcher = $dispatcher;
 		$this->language = $language;
-		$this->path_helper = $path_helper;
 		$this->ppde_actions = $ppde_actions;
 		$this->ppde_controller_main = $ppde_controller_main;
 		$this->ppde_controller_transactions_admin = $ppde_controller_transactions_admin;
 		$this->ppde_ipn_log = $ppde_ipn_log;
 		$this->ppde_ipn_paypal = $ppde_ipn_paypal;
 		$this->request = $request;
-		$this->php_ext = $php_ext;
-
-		$this->root_path = $this->path_helper->get_phpbb_root_path();
 	}
 
 	public function handle()
@@ -272,6 +257,7 @@ class ipn_listener
 		// Do actions only if checks are validated.
 		if ($this->validate_actions())
 		{
+			$this->prepare_data();
 			$this->do_actions();
 		}
 
@@ -544,6 +530,21 @@ class ipn_listener
 	}
 
 	/**
+	 * Some work to do before doing actions.
+	 *
+	 * @return void
+	 * @access private
+	 */
+	private function prepare_data()
+	{
+		$this->ppde_actions->set_transaction_data($this->transaction_data);
+		$this->ppde_actions->set_ipn_test_properties((bool) $this->transaction_data['test_ipn']);
+		$this->ppde_actions->is_donor_is_member();
+		$this->donor_is_member = $this->ppde_actions->get_donor_is_member();
+		$this->payer_data = $this->ppde_actions->get_payer_data();
+	}
+
+	/**
 	 * Validates actions if the transaction is verified
 	 *
 	 * @return bool
@@ -557,8 +558,8 @@ class ipn_listener
 			return false;
 		}
 
-		$this->tasks_list['payment_completed'] = $validate[] = $this->payment_status_is_completed();
-		$this->tasks_list['donor_is_member'] = $this->donor_is_member();
+		$this->tasks_list['payment_completed'] = $validate[] = $this->ppde_actions->payment_status_is_completed();
+		$this->tasks_list['donor_is_member'] = $this->donor_is_member;
 		$this->tasks_list['txn_errors'] = !empty($this->transaction_data['txn_errors']) && empty($this->transaction_data['txn_errors_approved']) ? true : false;
 		$this->tasks_list['is_not_ipn_test'] = !$this->transaction_data['test_ipn'];
 
@@ -593,9 +594,11 @@ class ipn_listener
 			unset($transaction_data);
 
 			// Do actions whether the transaction is real or a test.
-			$this->ppde_actions->set_transaction_data($this->transaction_data);
-			$this->ppde_actions->set_ipn_test_properties((bool) $this->transaction_data['test_ipn']);
 			$this->ppde_actions->update_overview_stats();
+			if (!$this->tasks_list['txn_errors'])
+			{
+				$this->ppde_actions->update_raised_amount();
+			}
 		}
 
 		if ($this->tasks_list['txn_errors'])
@@ -606,181 +609,15 @@ class ipn_listener
 
 		if ($this->tasks_list['is_not_ipn_test'])
 		{
-			$this->ppde_actions->update_raised_amount();
 			$this->ppde_actions->notification->notify_admin_donation_received();
 
 			if ($this->tasks_list['donor_is_member'])
 			{
-				$this->update_donor_stats();
-				$this->donors_group_user_add();
+				$this->ppde_actions->update_donor_stats();
+				$this->ppde_actions->donors_group_user_add();
 				$this->ppde_actions->notification->notify_donor_donation_received();
 			}
 		}
-	}
-
-	/**
-	 * Checks if payment_status is completed
-	 *
-	 * @return bool
-	 * @access private
-	 */
-	private function payment_status_is_completed()
-	{
-		return $this->transaction_data['payment_status'] === 'Completed';
-	}
-
-	/**
-	 * Returns if donor is member
-	 *
-	 * @return bool
-	 * @access private
-	 */
-	private function donor_is_member()
-	{
-		$this->is_donor_is_member();
-		return $this->donor_is_member && !empty($this->payer_data);
-	}
-
-	/**
-	 * Checks if the donor is a member then gets payer_data values
-	 *
-	 * @return void
-	 * @access private
-	 */
-
-	private function is_donor_is_member()
-	{
-		$anonymous_user = false;
-
-		// If the user_id is not anonymous
-		if ($this->transaction_data['user_id'] != ANONYMOUS)
-		{
-			$this->donor_is_member = $this->check_donors_status('user', $this->transaction_data['user_id']);
-
-			if (!$this->donor_is_member)
-			{
-				// No results, therefore the user is anonymous...
-				$anonymous_user = true;
-			}
-		}
-		else
-		{
-			// The user is anonymous by default
-			$anonymous_user = true;
-		}
-
-		if ($anonymous_user)
-		{
-			// If the user is anonymous, check their PayPal email address with all known email hashes
-			// to determine if the user exists in the database with that email
-			$this->donor_is_member = $this->check_donors_status('email', $this->transaction_data['payer_email']);
-		}
-	}
-
-	/**
-	 * Gets donor informations (user id, username, amount donated) and returns if exists
-	 *
-	 * @param string     $type Allowed value : 'user' or 'email'
-	 * @param string|int $args If $type is set to 'user', $args must be a user id.
-	 *                         If $type is set to 'email', $args must be an email address
-	 *
-	 * @return bool
-	 * @access private
-	 */
-	private function check_donors_status($type, $args)
-	{
-		$this->payer_data = $this->ppde_controller_transactions_admin->ppde_operator->query_donor_user_data($type, $args);
-
-		return !empty($this->payer_data);
-	}
-
-	private function update_donor_stats()
-	{
-		if ($this->donor_is_member)
-		{
-			$this->ppde_controller_transactions_admin->update_user_stats((int) $this->payer_data['user_id'], (float) $this->payer_data['user_ppde_donated_amount'] + (float) $this->ppde_actions->net_amount($this->transaction_data['mc_gross'], $this->transaction_data['mc_fee']));
-		}
-	}
-
-	/**
-	 * Add donor to the donors group
-	 *
-	 * @return void
-	 * @access private
-	 */
-	private function donors_group_user_add()
-	{
-		// We add the user to the donors group
-		$can_use_autogroup = $this->can_use_autogroup();
-		$group_id = (int) $this->config['ppde_ipn_group_id'];
-		$payer_id = (int) $this->payer_data['user_id'];
-		$payer_username = $this->payer_data['username'];
-		$default_group = $this->config['ppde_ipn_group_as_default'];
-
-		/**
-		 * Event to modify data before a user is added to the donors group
-		 *
-		 * @event skouat.ppde.donors_group_user_add_before
-		 * @var bool    can_use_autogroup   Whether or not to add the user to the group
-		 * @var int     group_id            The ID of the group to which the user will be added
-		 * @var int     payer_id            The ID of the user who will we added to the group
-		 * @var string  payer_username      The user name
-		 * @var bool    default_group       Whether or not the group should be made default for the user
-		 * @since 1.0.3
-		 */
-		$vars = array(
-			'can_use_autogroup',
-			'group_id',
-			'payer_id',
-			'payer_username',
-			'default_group',
-		);
-		extract($this->dispatcher->trigger_event('skouat.ppde.donors_group_user_add_before', compact($vars)));
-
-		if ($can_use_autogroup)
-		{
-			if (!function_exists('group_user_add'))
-			{
-				include($this->root_path . 'includes/functions_user.' . $this->php_ext);
-			}
-
-			// Adds the user to the donors group and set as default.
-			group_user_add($group_id, array($payer_id), array($payer_username), get_group_name($group_id), $default_group);
-		}
-	}
-
-	/**
-	 * Checks if all required settings are meet for adding the donor to the group of donors
-	 *
-	 * @return bool
-	 * @access private
-	 */
-	private function can_use_autogroup()
-	{
-		return
-			$this->autogroup_is_enabled() &&
-			$this->donor_is_member &&
-			$this->payment_status_is_completed() &&
-			$this->minimum_donation_raised();
-	}
-
-	/**
-	 * Checks if Autogroup could be used
-	 *
-	 * @return bool
-	 * @access private
-	 */
-	private function autogroup_is_enabled()
-	{
-		return $this->verified && $this->config['ppde_ipn_enable'] && $this->config['ppde_ipn_autogroup_enable'];
-	}
-
-	/**
-	 * @return bool
-	 */
-	private function minimum_donation_raised()
-	{
-		return (float) $this->payer_data['user_ppde_donated_amount'] >= (float) $this->config['ppde_ipn_min_before_group'] ? true : true;
 	}
 
 	/**
